refactor(itun): delete the share screen — sharing is a status on the sheet
The Share Snapshot screen's left half was a "Snapshot preview" that was not
what a recipient sees, and could not be. `SnapshotPreviewCard` hand-rolled a
hero band from `SheetHero` and re-derived every max-stat computation, then
rendered no rail, no systems, no cargo and no abilities — which is why it read
as raw and unfinished. Meanwhile `/s/:id` renders the REAL sheet read-only,
through `frozenSheet.ts` + `<Sheet readOnly />`, as `/p/:kind/:appId` and the
Game crew view already do.

Worse than incomplete, it disagreed: the preview deliberately showed the bare
entity while the published payload carries `context.pilotAbilities`, so a mech
previewed LOWER than it shared (Beefcake, ADR-029).

A better preview was never the fix. The sheet you are standing on IS the
preview, so sharing is a status you check, not a place you go:

  /sheet/:kind/:id/share  →  gone, route and all
  Share (top bar)         →  opens ShareStatusDialog over the live sheet

The dialog opens on the status — "Not shared", or "Shared — N live links" —
then publish, copy, QR, revoke, and the public-sheet toggle when connected.
Both share models stay together (ADR-004 frozen, ADR-032 live), because a
player choosing between them should see both.

Three things worth knowing:

- The probe is gated on `open`. The screen ran `probeSnapshotService` on mount,
  which was once per deliberate visit; this dialog is mounted by every live
  sheet, so an ungated probe would put a request on the app's most-visited
  surface for a feature most visits never touch. There is a test for it.
- Sections, not Panels. ModalShell already frames its content in a Card, and a
  bordered Panel inside a bordered Card is the double-framing the sheet rules
  forbid. `PublicSheetPanel` loses its Panel for the same reason — its only
  consumer is now this dialog.
- The QR survived, the QR panel did not. Passing a phone across a table is the
  actual use; it needs no heading and no "publish to generate a QR code"
  placeholder, so it is simply there once there is a link to encode.

Net −1,179 lines, one route, one screen and one duplicate of the derived-stat
math. `useEntity` stops being re-exported (its last consumer went with the
screen) — `usePilot`/`useMech` are the typed way in.

Verified in the running app at 1280 and 390, unpublished and published.
