+++ apps/itun/src/hooks/queries/index.ts
@@ -7,9 +7,12 @@
  * `useEntityStore.getState()` imperatively.
  */
 
+// `useEntity` is deliberately NOT re-exported: it is the generic these are
+// built on, and its last consumer went with the share screen. `usePilot` /
+// `useMech` are the typed way in — a caller wanting the generic is usually
+// about to lose the narrowing those give it.
 export {
   useCrawlers,
-  useEntity,
   useMech,
   useMechs,
   usePilot,
