+++ docs/design/itun-design-review.md
@@ -82,7 +82,7 @@ _Original finding text follows._
 - **T-3 — RESOLVED.** **Zero** hand-rolled dialogs remain. `UnassignLinkButton` and `DeleteConfirmDialog` were deleted outright; `AssignPilotToMech`, `AssignCrawlerToPilot`, `WorkspaceList`, and `SavePatternButton` all import `ModalShell`; `WorkspaceSwitcher` no longer contains a dialog at all. `useDialogA11y` has **zero** hits repo-wide, and `ModalShell` has ~20 ITUN consumers. The only remaining `role="dialog"` string in ITUN is an assertion inside `dashboard/__tests__/DisplayPanel.test.tsx`.
 - **T-4 — RESOLVED.** **Zero** occurrences of `bg-background`, `text-muted-foreground`, or `text-foreground` remain in `apps/itun/src` (including `wiring/`). The sweep finished; the proposed lint rule was not added.
 - **T-5 — STILL OPEN, and the largest surviving item.** The five specific values called out (`text-[10.5px]`, `text-[7px]`, `tracking-[0.12em]`, `px-[26px]`, `lg:w-[196px]`) are gone, but **~30 non-test ITUN files still carry arbitrary bracket values**, and no `LAYOUT` constants module exists. The named files `Sheet.tsx`, `StorageManifest.tsx` are still among them; `Erow.tsx` and `WizShell.tsx` no longer exist under those names.
-- **T-6 — PARTIAL.** Three of the four shrank substantially: `Sheet.tsx` 784 → **253**, `CrawlerSheet.tsx` 622 → **385**, `PilotSheet.tsx` 754 → **670**. `ShareSnapshotScreen.tsx` **grew**, 500 → **565**. The opportunistic approach recommended here was followed.
+- **T-6 — PARTIAL.** Three of the four shrank substantially: `Sheet.tsx` 784 → **253**, `CrawlerSheet.tsx` 622 → **385**, `PilotSheet.tsx` 754 → **670**. `ShareSnapshotScreen.tsx` **grew**, 500 → **565**, and was then **deleted outright** — the share screen became `ShareStatusDialog` on the live sheet, which is the shape that made most of those lines unnecessary rather than smaller.
 - **T-7 — RESOLVED.** The proposed `src/hooks/queries/` layer exists (`entities.ts`, `workspaces.ts`, `useHydrateEntities.ts`, barrel `index.ts`) and is consumed (e.g. `usePilots()` in `AssignPilotToMech`). Direct subscriptions in non-test `.tsx` are down from 46+ to **3** `useEntityStore(...)` plus **12** `useEntityStore.getState()`.
 
 _Original finding text follows._
@@ -142,7 +142,7 @@ The reviewer's structural verdict: leave `LiveSheet`/`SheetHero`, the hero↔str
 
 **Status: U-1, U-2, U-4, U-5 RESOLVED; U-3 SUPERSEDED; U-6 mostly RESOLVED** (verified 2026-07-21):
 
-- **U-1 — RESOLVED.** `packages/component-lib/src/components/stat/heatLevel.ts` defines `heatDangerFrom(max)` at a 0.7 ratio, citing this finding by number; `pipRows.ts` `trackSegmentState` returns a `'danger'` state past `dangerFrom`, and `VitalGauge.tsx` consumes it. ITUN passes it at `sheet/MechSheet.tsx:502` and `sheet/ShareSnapshotScreen.tsx:524`. Note `PIP_FILL` no longer exists as a symbol. The `--animate-heat-pulse` keyframe is defined in `theme.css` but **no component applies it** — the color escalation shipped, the pulse did not.
+- **U-1 — RESOLVED.** `packages/component-lib/src/components/stat/heatLevel.ts` defines `heatDangerFrom(max)` at a 0.7 ratio, citing this finding by number; `pipRows.ts` `trackSegmentState` returns a `'danger'` state past `dangerFrom`, and `VitalGauge.tsx` consumes it. ITUN passes it at `sheet/MechSheet.tsx:502` (the second call site, `ShareSnapshotScreen.tsx`, went with that screen). Note `PIP_FILL` no longer exists as a symbol. The `--animate-heat-pulse` keyframe is defined in `theme.css` but **no component applies it** — the color escalation shipped, the pulse did not.
 - **U-2 — RESOLVED.** `shared/GameDataReady.tsx` renders a branded `GameDataFallback`: SU cargo mark on `bg-ink-deep`, `font-cond uppercase` status text kept in the a11y tree via `role="status"`, and a `motion-safe:`-guarded rust loader bar.
 - **U-3 — SUPERSEDED, not built as specified.** There is no sheet FAB component (only a stale layout comment at `sheet/LiveSheet.tsx:272`). The function this finding wanted — banded d20 results with push bookkeeping — instead landed on the Dashboard via R-6 (`dashboard/ActionsDeck.tsx` + `CORE_ROLL_BANDS`), consistent with ADR-021 putting play actions on the Dashboard rather than the live sheet.
 - **U-4 — RESOLVED.** `roster/Roster.tsx` renders each `↳ Name` segment as a link to the target entity's live sheet (see the comment at `Roster.tsx:145` citing this finding, and the covering test in `roster/__tests__/Roster.test.tsx:303`). Note `EntityListItem` no longer exists; the Roster replaced it.
@@ -166,7 +166,7 @@ _Original finding text follows._
 
 > **⚠️ This roadmap is spent. Phases 1–4 are done and Phase 5 is mostly done — do not execute it.**
 >
-> What remains from the whole sequence, per the markers above: **T-5** (typography/spacing scale + `LAYOUT` module — untouched), **T-2** (`GenericSelectorDialog` — not built), the leftover half of **T-6** (`ShareSnapshotScreen.tsx`, which grew), and the **UI surfaces for R-3 and R-7**, whose rules layers are written and tested but have zero component consumers. Everything else in Phases 1–5 shipped, though R-1/R-6/U-3 landed on the Dashboard rather than on the live sheets.
+> What remains from the whole sequence, per the markers above: **T-5** (typography/spacing scale + `LAYOUT` module — untouched), **T-2** (`GenericSelectorDialog` — not built), the leftover half of **T-6** (`ShareSnapshotScreen.tsx` — since resolved by deletion, not extraction), and the **UI surfaces for R-3 and R-7**, whose rules layers are written and tested but have zero component consumers. Everything else in Phases 1–5 shipped, though R-1/R-6/U-3 landed on the Dashboard rather than on the live sheets.
 
 Sequenced so each phase is independently shippable and the tidiness work lands _before_ the features that would otherwise copy the mess.
